to string method finished and passing all tests added error handling to each mehtod as well
