+++ python/code_challenges/challenge_05A/README.md
@@ -1,0 +1,12 @@
+# Singly Linked List
+<!-- Short summary or background information -->
+
+## Challenge
+<!-- Description of the challenge -->
+
+## Approach & Efficiency
+<!-- What approach did you take? Why? What is the Big O space/time for this approach? -->
+
+## API
+<!-- Description of each method publicly available to your Linked List -->
+
