--- python/tests/test_reverse_linked_list.py
@@ -1,5 +1,0 @@
-from code_challenges.reverse_linked_list import reverse_list
-
-
-def test_import():
-    assert reverse_list
